pyproject.toml
[project]
name = "diffumoji"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cairosvg>=2.8.2",
    "datasets>=4.1.0",
    "ema-pytorch>=0.7.7",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "matplotlib>=3.10.6",
    "pillow>=11.3.0",
    "torch",
    "torchaudio",
    "torchvision",
    "transformers>=4.56.1",
    "wandb[media]>=0.22.0",
]
